+++ api/Ourchitecture.Api/Ourchitecture.Api.csproj
@@ -299,6 +299,7 @@
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Routes\BaseModule.cs" />
+    <Compile Include="Routes\GrasshopperModule.cs" />
     <Compile Include="Startup.cs" />
   </ItemGroup>
   <ItemGroup>
