+++ public/src/app/app-routing.module.ts
@@ -3,9 +3,11 @@ import { Routes, RouterModule } from '@angular/router';
 
 import { RootComponent } from './root/root.component';
 import { CiceroComponent } from './cicero/cicero.component';
+import { SystemsComponent } from './systems/systems.component';
 
 const routes: Routes = [
-  { path: "", redirectTo: "/root", pathMatch: "full" },
+  { path: "", redirectTo: "/systems", pathMatch: "full" },
+  { path: "systems", component: SystemsComponent},
   { path: "root", component: RootComponent},
   { path: "root/test", component: RootComponent},
   { path: "cicero", component: CiceroComponent}
