+++ public/src/app/app.component.css
@@ -1,65 +1,19 @@
 .wrapper_main {
     display: grid;
-    grid-template-rows: 7vh 1fr;
-    grid-template-columns: 11vh 1fr 7vh;
-}
-
-.header {
-    grid-row: 1 / span 1;
-    grid-column: 1 / span 3;
-
-    display: grid;
-    grid-template-rows: 1fr 1fr;
-    grid-template-columns: 1fr 7vh;
-
-    background: none;
-
-    border-bottom: 4px solid black;
-
-    overflow: hidden;
-}
-
-.icon {
-    grid-row: 1 / span 2;
-    grid-column: 2 / span 1;
-
-    border-left: 4px solid black;
-}
-
-.icon:hover {
-    cursor: pointer;
-    background: black;
-}
-
-.active_header {
-    width: 200vw;
-    background: black;
-    transform: translateX(-100vw);
-
-    animation-name: activate-header;
-    animation-duration: 1s;
-}
-
-@keyframes activate-header {
-    0%   {transform: translateX(0)}
-    100% {transform: translateX(-100vw)}
-}
-
-.margin {
-    grid-row: 2 / span 1;
-    grid-column: 1 / span 1;
-
-    border-right: 4px dashed black;
-
-    background: none;
-    pointer-events: none;
+    grid-template-rows: 10vw calc(100vh - 20vw) 10vw;
+    grid-template-columns: 10vw 80vw 10vw;
 }
 
 .content {
     grid-row: 2 / span 1;
+    grid-column: 2 / span 1;
+
+    background: yellow;
+}
+
+.navigation {
+    grid-row: 3 / span 1;
     grid-column: 1 / span 3;
-
-    height: 93vh;
-
-    overflow-y: auto;
+    
+    background: green;
 }
