+++ public/src/app/app.component.html
@@ -1,15 +1,14 @@
 <div class="wrapper_main">
 
-  <div class="header">
-    <div class="title"></div>
-    <div class="icon" id="header" (click)="onActivateHeader()" [ngClass]="{'active_header': headerActive === true}"></div>
-  </div>
-
   <div class="content">
       <router-outlet></router-outlet>
   </div>
 
-  <div class="margin"></div>
+  <div class="navigation">
+    <div class="navigation__left"></div>
+    <div class="navigation__title"></div>
+    <div class="navigation__right"></div>
+  </div>
 
 </div>
