+++ public/src/app/systems/systems.component.ts
@@ -1,0 +1,15 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'app-systems',
+  templateUrl: './systems.component.html',
+  styleUrls: ['./systems.component.css']
+})
+export class SystemsComponent implements OnInit {
+
+  constructor() { }
+
+  ngOnInit() {
+  }
+
+}
